updated customer and till classes and tests for buying vehicles

+++ src/main/java/vehicles/Car.java
@@ -5,7 +5,7 @@ import Components.IComponents;
 import java.util.ArrayList;
 
 public class Car extends Vehicle {
-    public Car(int price, String colour, ArrayList<IComponents> components) {
+    public Car(double price, String colour, ArrayList<IComponents> components) {
         super(price, colour, components);
     }
 }
