+++ src/main/java/vehicles/ElectricCar.java
@@ -5,7 +5,7 @@ import Components.IComponents;
 import java.util.ArrayList;
 
 public class ElectricCar extends Vehicle {
-    public ElectricCar(int price, String colour, ArrayList<IComponents> components) {
+    public ElectricCar(double price, String colour, ArrayList<IComponents> components) {
         super(price, colour, components);
     }
 }
