+++ src/main/java/vehicles/HybridCar.java
@@ -5,7 +5,7 @@ import Components.IComponents;
 import java.util.ArrayList;
 
 public class HybridCar extends Vehicle {
-    public HybridCar(int price, String colour, ArrayList<IComponents> components) {
+    public HybridCar(double price, String colour, ArrayList<IComponents> components) {
         super(price, colour, components);
     }
 }
